Refactored models into own files, updated logic to carry all pertinent info inside the CallDetails List Object instead of a Dictionary Object. Decoupling output from Calulcatation logic to add further output options/extensions .
