+++ models/PlanDetails.cs
@@ -1,0 +1,11 @@
+namespace callRecords.Models
+{
+    /// <summary>
+    /// Plan limit to Friendly Name Mapping POCO
+    /// </summary>
+    public class PlanDetails
+    {
+        public string planTypeFriendlyName { get; set; }
+        public int planLimit { get; set; }
+    }
+}
